Web: Broken state for deleting branch in repo that deletes on merge Fixes #3453
